+++ parser_test.go
@@ -3256,312 +3256,6 @@ func TestDefineTypeOfExample(t *testing.T) {
 	})
 }
 
-func TestParseFieldTag(t *testing.T) {
-
-	t.Run("Example tag", func(t *testing.T) {
-		t.Parallel()
-
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" example:"one"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		assert.Equal(t, &structField{
-			schemaType:   "string",
-			exampleValue: "one",
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" example:"one"`,
-				},
-			},
-			[]string{"float"})
-		assert.Error(t, err)
-	})
-
-	t.Run("Format tag", func(t *testing.T) {
-		t.Parallel()
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" format:"csv"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		assert.Equal(t, &structField{
-			schemaType: "string",
-			formatType: "csv",
-		}, field)
-	})
-
-	t.Run("Required tag", func(t *testing.T) {
-		t.Parallel()
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" binding:"required"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		assert.Equal(t, &structField{
-			schemaType: "string",
-			isRequired: true,
-		}, field)
-
-		field, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" validate:"required"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		assert.Equal(t, &structField{
-			schemaType: "string",
-			isRequired: true,
-		}, field)
-
-	})
-
-	t.Run("Extensions tag", func(t *testing.T) {
-
-	})
-
-	t.Run("Enums tag", func(t *testing.T) {
-		t.Parallel()
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" enums:"a,b,c"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		assert.Equal(t, &structField{
-			schemaType: "string",
-			enums:      []interface{}{"a", "b", "c"},
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" enums:"a,b,c"`,
-				},
-			},
-			[]string{"float"})
-		assert.Error(t, err)
-	})
-
-	t.Run("Default tag", func(t *testing.T) {
-		t.Parallel()
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" default:"pass"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		assert.Equal(t, &structField{
-			schemaType:   "string",
-			defaultValue: "pass",
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" default:"pass"`,
-				},
-			},
-			[]string{"float"})
-		assert.Error(t, err)
-	})
-
-	t.Run("Numeric value", func(t *testing.T) {
-		t.Parallel()
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" maximum:"1"`,
-				},
-			},
-			[]string{"integer"})
-		assert.NoError(t, err)
-		max := float64(1)
-		assert.Equal(t, &structField{
-			schemaType: "integer",
-			maximum:    &max,
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" maximum:"one"`,
-				},
-			},
-			[]string{"integer"})
-		assert.Error(t, err)
-
-		field, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" maximum:"1"`,
-				},
-			},
-			[]string{"number"})
-		assert.NoError(t, err)
-		max = float64(1)
-		assert.Equal(t, &structField{
-			schemaType: "number",
-			maximum:    &max,
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" maximum:"one"`,
-				},
-			},
-			[]string{"number"})
-		assert.Error(t, err)
-
-		field, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" multipleOf:"1"`,
-				},
-			},
-			[]string{"number"})
-		assert.NoError(t, err)
-		multipleOf := float64(1)
-		assert.Equal(t, &structField{
-			schemaType: "number",
-			multipleOf: &multipleOf,
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" multipleOf:"one"`,
-				},
-			},
-			[]string{"number"})
-		assert.Error(t, err)
-
-		field, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" minimum:"1"`,
-				},
-			},
-			[]string{"integer"})
-		assert.NoError(t, err)
-		min := float64(1)
-		assert.Equal(t, &structField{
-			schemaType: "integer",
-			minimum:    &min,
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" minimum:"one"`,
-				},
-			},
-			[]string{"integer"})
-		assert.Error(t, err)
-
-	})
-
-	t.Run("String value", func(t *testing.T) {
-		t.Parallel()
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" maxLength:"1"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		max := int64(1)
-		assert.Equal(t, &structField{
-			schemaType: "string",
-			maxLength:  &max,
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" maxLength:"one"`,
-				},
-			},
-			[]string{"string"})
-		assert.Error(t, err)
-
-		field, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" minLength:"1"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		min := int64(1)
-		assert.Equal(t, &structField{
-			schemaType: "string",
-			minLength:  &min,
-		}, field)
-
-		_, err = parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" minLength:"one"`,
-				},
-			},
-			[]string{"string"})
-		assert.Error(t, err)
-	})
-
-	t.Run("Readonly tag", func(t *testing.T) {
-		t.Parallel()
-		parser := New()
-
-		field, err := parser.parseFieldTag(
-			&ast.Field{
-				Tag: &ast.BasicLit{
-					Value: `json:"test" readonly:"true"`,
-				},
-			},
-			[]string{"string"})
-		assert.NoError(t, err)
-		assert.Equal(t, &structField{
-			schemaType: "string",
-			readOnly:   true,
-		}, field)
-	})
-}
-
 func TestSetRouteMethodOp(t *testing.T) {
 	t.Parallel()
 
@@ -3657,9 +3351,9 @@ func TestParser_Skip(t *testing.T) {
 
 	parser = New(SetExcludedDirsAndFiles("admin/release,admin/models"))
 	assert.NoError(t, parser.Skip("admin", &mockFS{IsDirectory: true}))
-	assert.NoError(t, parser.Skip("admin/service", &mockFS{IsDirectory: true}))
-	assert.Error(t, parser.Skip("admin/models", &mockFS{IsDirectory: true}))
-	assert.Error(t, parser.Skip("admin/release", &mockFS{IsDirectory: true}))
+	assert.NoError(t, parser.Skip(filepath.Clean("admin/service"), &mockFS{IsDirectory: true}))
+	assert.Error(t, parser.Skip(filepath.Clean("admin/models"), &mockFS{IsDirectory: true}))
+	assert.Error(t, parser.Skip(filepath.Clean("admin/release"), &mockFS{IsDirectory: true}))
 }
 
 func TestGetFieldType(t *testing.T) {
